requirements.txt
uvicorn==0.22.0
mysql-connector-python==8.0.33
fastapi==0.103.2
gunicorn==21.2.0
pydantic==2.4.2
pydantic_core==2.10.1